# docs/openapi-subset.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LanPro API (#315 gelombang 3 — dari Zod)",
    "version": "0.3.0",
    "description": "Digenerate dari skema Zod (`server/schemas`) + pemindaian validasiBody/Query/Params di `server/routes`. Prefix kanonik `/api`; `/api/v1` legacy (file/meeting AI). Envelope error: { status, code, message, errors? }. Tanpa Swagger UI. Idempotency-Key belum."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
  "paths": {
...
    "/api/v1/files/secure-stream": {
      "get": {
        "summary": "GET /api/v1/files/secure-stream",
        "tags": ["files"],
        "x-source": "server/routes/file.routes.ts",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "file",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            }
          },
          {
            "name": "expires",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 2048
            }
          },
          {
            "name": "uid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ]
      }
    },
...
    "/api/projects/{projectId}/milestones": {
      "get": {
        "summary": "GET /api/projects/:projectId/milestones",
        "tags": ["projects"],
        "x-source": "server/routes/milestones.routes.ts",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createMilestoneSchema"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "POST /api/projects/:projectId/milestones",
        "tags": ["projects"],
        "x-source": "server/routes/milestones.routes.ts",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createMilestoneSchema"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/projects/{projectId}/milestones/{id}": {
      "put": {
        "summary": "PUT /api/projects/:projectId/milestones/:id",
        "tags": ["projects"],
        "x-source": "server/routes/milestones.routes.ts",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateMilestoneSchema"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": ["status", "code", "message"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["error"]
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "createMilestoneSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ]
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ]
          },
          "sprintIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["name"],
        "additionalProperties": false
      },
      "updateMilestoneSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ]
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ]
          },
          "sprintIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}